Breaking NodeGeneralForm into DescriptionField and NameField for reuse
